Long -> BigDecimal

+++ sources/repository/repository-actions/src/main/kotlin/app/actionsfun/repository/actions/interactor/DepositToMarketInteractor.kt
@@ -4,10 +4,11 @@ import app.actionsfun.repository.actions.interactor.tx.CreateTransactionInteract
 import app.actionsfun.repository.solana.WalletRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import java.math.BigDecimal
 
 interface DepositToMarketInteractor {
 
-    suspend fun deposit(marketAddress: String, amount: Long, option: Boolean): String
+    suspend fun deposit(marketAddress: String, amount: BigDecimal, option: Boolean): String
 }
 
 internal class DepositToMarketInteractorImpl(
@@ -15,7 +16,7 @@ internal class DepositToMarketInteractorImpl(
     private val walletRepository: WalletRepository,
 ) : DepositToMarketInteractor {
 
-    override suspend fun deposit(marketAddress: String, amount: Long, option: Boolean): String {
+    override suspend fun deposit(marketAddress: String, amount: BigDecimal, option: Boolean): String {
         return withContext(Dispatchers.IO) {
             val transaction = createTransactionInteractor.createDepositToMarketTransaction(
                 marketAddress = marketAddress,
